// src/types/GreekModes/GreekModeType.ts
export enum GreekModeType {
  Ionian = "Ionian",
  Dorian = "Dorian",
  UkrainianDorian = "Ukrainian",
  Phrygian = "Phrygian",
  Spanish = "Spanish", //aka Phrygian Dominant
  Arabic = "Arabic",
  Lydian = "Lydian",
  Mixolydian = "Mixolydian",
  Aeolian = "Aeolian",
  HarmonicMinor = "Harmonic",
  HungarianMinor = "Hungarian",
  Locrian = "Locrian",
}